drsalam/requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-stdimage==5.1.1
djangorestframework==3.12.1
Khayyam==3.0.17
Pillow==8.0.1
progressbar2==3.53.1
python-utils==2.4.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
